Moved the query mechanism into command device from device. Only command should be used to send queries.
