+++ Devices/igtlioCommandDevice.h
@@ -13,6 +13,16 @@ typedef vtkSmartPointer<class CommandDevice> CommandDevicePointer;
 class OPENIGTLINKIO_DEVICES_EXPORT CommandDevice : public Device
 {
 public:
+	enum QUERY_STATUS {
+	  QUERY_STATUS_NONE,
+	  QUERY_STATUS_WAITING,
+	  QUERY_STATUS_SUCCESS,
+	  QUERY_STATUS_EXPIRED,
+	  QUERY_STATUS_ERROR,
+	  NUM_QUERY_STATUS,
+	};
+
+public:
  virtual std::string GetDeviceType() const;
  virtual int ReceiveIGTLMessage(igtl::MessageBase::Pointer buffer, bool checkCRC);
  virtual igtl::MessageBase::Pointer GetIGTLMessage();
@@ -26,6 +36,39 @@ public:
   igtl::MessageBase::Pointer GetIGTLResponseMessage();
   CommandDevicePointer GetResponseFromCommandID(int id);
 
+  /// Query handling:
+  /// Each device has a list of queries (GET_, STT_, STP_) that has been sent
+  /// and are awaiting reply.
+  //
+  /// Device::GetMessage() pushes a query,
+  /// Device::ReceiveMessage() processes the reply, and emits events for the receive
+  ///   - statechange: waiting, success, expired,...
+  ///
+  /// One query (GET_, STT_ or STP_-message that requires an answer)
+  ///
+  /// TODO: Currently implemented for COMMAND message only. The GET_/STT_/STP_ messages
+  /// handle this by simply sending and ignoring failures.
+  /// Either move the query mechanism down to COMMAND or generalize.
+  ///
+  ///
+  struct QueryType
+  {
+	DevicePointer Query;
+	DevicePointer Response;
+	QUERY_STATUS status;
+  };
+
+   /// Get all current queries
+   std::vector<QueryType> GetQueries() const;
+
+   /// check for waiting queries that have waited beoynd the timeout for an answer, mark them as expired.
+
+   int CheckQueryExpiration();
+   /// remove all queries that are answered or expired.
+   int PruneCompletedQueries();
+
+   int CancelQuery(int index);
+
  public:
   static CommandDevice *New();
   vtkTypeMacro(CommandDevice,Device);
@@ -37,6 +80,7 @@ public:
   ~CommandDevice();
 
  protected:
+  std::vector<QueryType> Queries;
   igtl::CommandMessage::Pointer OutMessage;
   igtl::RTSCommandMessage::Pointer ResponseMessage;
   CommandConverter::ContentData Content;
