+++ Devices/igtlioDevice.h
@@ -50,14 +50,7 @@ public:
    MESSAGE_PREFIX_RTS,
    NUM_MESSAGE_PREFIX,
  };
- enum QUERY_STATUS {
-   QUERY_STATUS_NONE,
-   QUERY_STATUS_WAITING,
-   QUERY_STATUS_SUCCESS,
-   QUERY_STATUS_EXPIRED,
-   QUERY_STATUS_ERROR,
-   NUM_QUERY_STATUS,
- };
+
  enum {
    ReceiveEvent          = 118948,
    ResponseEvent         = 118952,
@@ -116,38 +109,6 @@ public:
  //       - lock (means dont accept incoming messages),
  //       - gettimestamp (of last incoming message)
 
- /// Query handling:
- /// Each device has a list of queries (GET_, STT_, STP_) that has been sent
- /// and are awaiting reply.
- //
- /// Device::GetMessage() pushes a query,
- /// Device::ReceiveMessage() processes the reply, and emits events for the receive
- ///   - statechange: waiting, success, expired,...
- ///
- /// One query (GET_, STT_ or STP_-message that requires an answer)
- ///
- /// TODO: Currently implemented for COMMAND message only. The GET_/STT_/STP_ messages
- /// handle this by simply sending and ignoring failures.
- /// Either move the query mechanism down to COMMAND or generalize.
- ///
- ///
- struct QueryType
- {
-   DevicePointer Query;
-   DevicePointer Response;
-   QUERY_STATUS status;
- };
-
-  /// Get all current queries
-  std::vector<QueryType> GetQueries() const;
-
-  /// check for waiting queries that have waited beoynd the timeout for an answer, mark them as expired.
-
-  int CheckQueryExpiration();
-  /// remove all queries that are answered or expired.
-  int PruneCompletedQueries();
-
-  int CancelQuery(int index);
 
 public:
   vtkAbstractTypeMacro(Device,vtkObject);
@@ -155,7 +116,6 @@ public:
 protected:
   void SetHeader(BaseConverter::HeaderData header);
 
-  std::vector<QueryType> Queries;
   BaseConverter::HeaderData HeaderData;
 
 protected:
